+++ flutter_navigation_generator/example/lib/main.dart
@@ -37,13 +37,13 @@ class MyApp extends StatelessWidget {
 )
 @FlutterRoute(
   methodName: 'goToHomePageWithPathParameter',
-  routeName: 'my-home-page-pop-all/:title',
+  routeName: 'my-home-page-pop-all/',
   navigationType: NavigationType.pushAndReplaceAll,
 )
 class MyHomePage extends StatefulWidget {
   const MyHomePage({
-    super.key,
     this.title,
+    @FlutterRouteField(ignore: false) super.key, // Key is ignored by default
   });
 
   @FlutterRouteConstructor(
@@ -51,6 +51,7 @@ class MyHomePage extends StatefulWidget {
   )
   const MyHomePage.popAll({super.key, this.title = 'Popped all pages'});
 
+  @FlutterRouteField(queryName: 'page_title')
   final String? title;
 
   @override
@@ -105,14 +106,11 @@ class _MyHomePageState extends State<MyHomePage> {
               child: const Text("Go to page 2 (with fade animation)"),
             ),
             ElevatedButton(
-              onPressed: () => mainNavigator
-                  .goToExampleScreenWithRequiredArgument(
-                      data: [CustomModel('John', 25), CustomModel('Jeff', 27)]),
+              onPressed: () => mainNavigator.goToExampleScreenWithRequiredArgument(data: [CustomModel('John', 25), CustomModel('Jeff', 27)]),
               child: const Text("Go to ExampleScreenWithRequiredArgument"),
             ),
             ElevatedButton(
-              onPressed: () =>
-                  mainNavigator.customName(id: '1', name: 'John', age: 12),
+              onPressed: () => mainNavigator.customName(id: '1', name: 'John', age: 12),
               child: const Text("Go to RouteNameWithArguments"),
             ),
             ElevatedButton(
@@ -132,17 +130,14 @@ class _MyHomePageState extends State<MyHomePage> {
               child: const Text("Go to RouteNameWithArguments2"),
             ),
             ElevatedButton(
-              onPressed: () =>
-                  mainNavigator.showSheetRecursiveNavigationBottomSheet(),
+              onPressed: () => mainNavigator.showSheetRecursiveNavigationBottomSheet(),
               child: const Text("Show a bottom sheet with its own navigator"),
             ),
             ElevatedButton(
-              onPressed: () =>
-                  mainNavigator.showDialogExampleDialog(text: 'hi there'),
+              onPressed: () => mainNavigator.showDialogExampleDialog(text: 'hi there'),
               child: const Text("Show a full screen dialog"),
             ),
-            Text(
-                "Has a navigation blocked by a guard (not logged in): ${mainNavigator.canContinueNavigation()}"),
+            Text("Has a navigation blocked by a guard (not logged in): ${mainNavigator.canContinueNavigation()}"),
             if (mainNavigator.canContinueNavigation()) ...[
               ElevatedButton(
                 onPressed: () async {
@@ -365,9 +360,7 @@ class RecursiveNavigationBottomSheet extends StatelessWidget {
                 child: const Text("Go to second page"),
               ),
               ElevatedButton(
-                onPressed: () =>
-                    myNavigator.showSheetRecursiveNavigationBottomSheet(
-                        layers: layers + 1),
+                onPressed: () => myNavigator.showSheetRecursiveNavigationBottomSheet(layers: layers + 1),
                 child: const Text("Open another bottom sheet"),
               ),
             ],
@@ -472,8 +465,7 @@ class Error404 extends StatelessWidget {
               'We couldn\'t find this page, sorry :(',
             ),
             ElevatedButton(
-              onPressed: () =>
-                  mainNavigator.goToMyHomePage(title: 'returning from 404'),
+              onPressed: () => mainNavigator.goToMyHomePage(title: 'returning from 404'),
               child: const Text("go home"),
             ),
           ],
@@ -506,8 +498,7 @@ class ErrorNotLoggedIn extends StatelessWidget {
               'You are not logged in, sorry :(',
             ),
             ElevatedButton(
-              onPressed: () => mainNavigator.goToMyHomePage(
-                  title: 'returning from not logged in'),
+              onPressed: () => mainNavigator.goToMyHomePage(title: 'returning from not logged in'),
               child: const Text("go home"),
             ),
             ElevatedButton(
@@ -550,8 +541,7 @@ class LoggedInPage extends StatelessWidget {
               'You are logged in, yay :)',
             ),
             ElevatedButton(
-              onPressed: () => mainNavigator.goToMyHomePage(
-                  title: 'returning from logged in'),
+              onPressed: () => mainNavigator.goToMyHomePage(title: 'returning from logged in'),
               child: const Text("go home"),
             ),
           ],