+++ flutter_navigation_generator/lib/src/code_builder/on_generate_route_builder.dart
@@ -33,30 +33,26 @@ class OnGenerateRouteBuilder {
   }
 
   String _generateRoute(RouteConfig route) {
-    final constructor = route.constructorName == route.routeWidget.className ||
-            route.constructorName.isEmpty
+    final constructor = route.constructorName == route.routeWidget.className || route.constructorName.isEmpty
         ? route.routeWidget.className
         : '${route.routeWidget.className}.${route.constructorName}';
-    final constructorCall =
-        '$constructor(${route.parameters.asMap().map((_, p) {
-              final nullableSuffix = p.isNullable ? '?' : '';
-              final convertFromString =
-                  ImportableTypeStringConverter.convertFromString(
-                      p, 'arguments[\'${p.argumentName}\']');
-              return MapEntry(
-                p.argumentName,
-                (p.className == 'Key' || p.className == 'String')
-                    ? "arguments['${p.argumentName}'] as ${p.className}$nullableSuffix"
-                    : "arguments['${p.argumentName}'] is String ? $convertFromString : arguments['${p.argumentName}'] as ${p.className}${p.typeArguments.isNotEmpty ? '<${p.typeArguments.map((e) => e.className).join(',')}>' : ''}$nullableSuffix",
-              );
-            }).entries.map((e) => '${e.key}: ${e.value},').join('')})';
+    final constructorCall = '$constructor(${route.parameters.asMap().map((_, parameterConfig) {
+          final parameter = parameterConfig.type;
+          final nullableSuffix = parameter.isNullable ? '?' : '';
+          final convertFromString = ImportableTypeStringConverter.convertFromString(parameter, 'arguments[\'${parameter.argumentName}\']');
+          return MapEntry(
+            parameter.argumentName,
+            (parameter.className == 'Key' || parameter.className == 'String')
+                ? "arguments['${parameter.argumentName}'] as ${parameter.className}$nullableSuffix"
+                : "arguments['${parameter.argumentName}'] is String ? $convertFromString : arguments['${parameter.argumentName}'] as ${parameter.className}${parameter.typeArguments.isNotEmpty ? '<${parameter.typeArguments.map((e) => e.className).join(',')}>' : ''}$nullableSuffix",
+          );
+        }).entries.map((e) => '${e.key}: ${e.value},').join('')})';
 
     var guardsCode = '';
     if ((route.guards ?? defaultGuards).isNotEmpty) {
       for (final guard in route.guards ?? defaultGuards) {
         final insanceName = CaseUtil(guard.className).camelCase;
-        guardsCode +=
-            'final $insanceName = guards.whereType<${typeRefer(guard).symbol}>().first;\n';
+        guardsCode += 'final $insanceName = guards.whereType<${typeRefer(guard).symbol}>().first;\n';
         guardsCode += 'if (!$insanceName.value) {\n';
         guardsCode += '  guardedRouteSettings = settings;\n';
         guardsCode += '  return onGenerateRoute(RouteSettings(\n';
@@ -70,15 +66,9 @@ class OnGenerateRouteBuilder {
   }
 
   Method generate() {
-    final pageRoutes = routes
-        .where((r) =>
-            r.generatePageRoute &&
-            r.navigationType != NavigationType.bottomSheet &&
-            r.navigationType != NavigationType.dialog)
-        .toList();
+    final pageRoutes = routes.where((r) => r.generatePageRoute && r.navigationType != NavigationType.bottomSheet && r.navigationType != NavigationType.dialog).toList();
     for (final pageRoute in pageRoutes.toList()) {
-      final pageRoute2 = pageRoutes
-          .firstWhere((element) => element.routeName == pageRoute.routeName);
+      final pageRoute2 = pageRoutes.firstWhere((element) => element.routeName == pageRoute.routeName);
       if (pageRoute != pageRoute2) {
         pageRoutes.remove(pageRoute);
       }
@@ -96,8 +86,7 @@ class OnGenerateRouteBuilder {
         )
         ..body = Block.of([
           if (pageRoutes.isNotEmpty) ...[
-            const Code(
-                '''final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};
+            const Code('''final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};
     final settingsUri = Uri.parse(settings.name ?? '');
     settingsUri.queryParameters.forEach((key, value) {
       arguments[key] ??= value;
@@ -105,38 +94,28 @@ class OnGenerateRouteBuilder {
             Code(
                 'switch (settingsUri.path) {${pageRoutes.where((route) => !route.routeNameContainsParameters).map((route) => 'case RouteNames.${route.asRouteName}: ${_generateRoute(route)}').join('')}}'),
           ],
-          if (pageRoutes
-              .any((element) => element.routeNameContainsParameters)) ...[
+          if (pageRoutes.any((element) => element.routeNameContainsParameters)) ...[
             const Code('final pathSegments = settingsUri.pathSegments;'),
             ...pageRoutes
                 .where((pageRoute) => pageRoute.routeNameContainsParameters)
-                .groupListsBy(
-                    (pageRoute) => pageRoute.routeName.pathSegments.length)
+                .groupListsBy((pageRoute) => pageRoute.routeName.pathSegments.length)
                 .entries
                 .sorted((a, b) => -a.key.compareTo(b.key))
                 .map((group) {
               final pathSegments = group.key;
               var code = 'if (pathSegments.length == $pathSegments) {';
-              final pageRoutesMap = group.value.asMap().map((key, value) =>
-                  MapEntry(
-                      value, value.routeName.parametersFromRouteName.length));
-              final pageRoutes = pageRoutesMap.entries
-                  .sorted((a, b) => a.value.compareTo(b.value))
-                  .map((e) => e.key)
-                  .toList();
+              final pageRoutesMap = group.value.asMap().map((key, value) => MapEntry(value, value.routeName.parametersFromRouteName.length));
+              final pageRoutes = pageRoutesMap.entries.sorted((a, b) => a.value.compareTo(b.value)).map((e) => e.key).toList();
               for (final pageRoute in pageRoutes) {
                 final pathSegments = pageRoute.routeName.pathSegments;
-                final hasRigidSegments =
-                    pathSegments.any((element) => !element.startsWith(':'));
+                final hasRigidSegments = pathSegments.any((element) => !element.startsWith(':'));
                 if (hasRigidSegments) {
                   code += 'if (';
                   code += pathSegments
                       .asMap()
                       .entries
-                      .where(
-                          (pathSegment) => !pathSegment.value.startsWith(':'))
-                      .map((pathSegment) =>
-                          'pathSegments[${pathSegment.key}] == \'${pathSegment.value}\'')
+                      .where((pathSegment) => !pathSegment.value.startsWith(':'))
+                      .map((pathSegment) => 'pathSegments[${pathSegment.key}] == \'${pathSegment.value}\'')
                       .join(' && ');
                   code += ') {';
                 }
@@ -144,8 +123,7 @@ class OnGenerateRouteBuilder {
                     .asMap()
                     .entries
                     .where((pathSegment) => pathSegment.value.startsWith(':'))
-                    .map((pathSegment) =>
-                        'arguments[\'${pathSegments[pathSegment.key].substring(1)}\'] = pathSegments[${pathSegment.key}];')
+                    .map((pathSegment) => 'arguments[\'${pathSegments[pathSegment.key].substring(1)}\'] = pathSegments[${pathSegment.key}];')
                     .join('\n');
                 code += _generateRoute(pageRoute);
                 if (hasRigidSegments) code += '}';
@@ -155,8 +133,7 @@ class OnGenerateRouteBuilder {
             }),
           ],
           if (unknownRoute != null) ...[
-            Code(
-                'return ${_withPageType(null, '${typeRefer(unknownRoute!).symbol!}()')};'),
+            Code('return ${_withPageType(null, '${typeRefer(unknownRoute!).symbol!}()')};'),
           ] else ...[
             const Code('return null;'),
           ],