+++ flutter_navigation_generator/lib/src/models/route_config.dart
@@ -1,6 +1,7 @@
 import 'dart:convert';
 
 import 'package:flutter_navigation_generator/src/models/importable_type.dart';
+import 'package:flutter_navigation_generator/src/models/route_field_config.dart';
 import 'package:flutter_navigation_generator_annotations/flutter_navigation_generator_annotations.dart';
 
 class RouteConfig {
@@ -16,9 +17,8 @@ class RouteConfig {
   final ImportableType? returnType;
   final ImportableType? pageType;
   final NavigationType navigationType;
-  final List<ImportableType> parameters;
+  final List<RouteFieldConfig> parameters;
   final List<ImportableType>? guards;
-  final Map<String, dynamic> defaultValues;
 
   bool get routeNameContainsParameters => routeName.contains(':');
 
@@ -37,7 +37,6 @@ class RouteConfig {
     this.navigationType = NavigationType.push,
     this.parameters = const [],
     this.guards,
-    this.defaultValues = const {},
   });
 
   Map<String, dynamic> toMap() {
@@ -56,7 +55,6 @@ class RouteConfig {
       'navigationType': navigationType.index,
       'parameters': parameters.map((x) => x.toMap()).toList(),
       'guards': guards?.map((x) => x.toMap()).toList(),
-      'defaultValues': defaultValues,
     };
   }
 
@@ -68,32 +66,18 @@ class RouteConfig {
       constructorName: map['constructorName'] ?? '',
       routeName: map['routeName'] ?? '',
       methodName: map['methodName'] ?? '',
-      routeNameIsDefinedByAnnotation:
-          map['routeNameIsDefinedByAnnotation'] ?? '',
-      methodNameIsDefinedByAnnotation:
-          map['methodNameIsDefinedByAnnotation'] ?? '',
+      routeNameIsDefinedByAnnotation: map['routeNameIsDefinedByAnnotation'] ?? '',
+      methodNameIsDefinedByAnnotation: map['methodNameIsDefinedByAnnotation'] ?? '',
       routeWidget: ImportableType.fromMap(map['routeWidget']),
-      returnType: map['returnType'] != null
-          ? ImportableType.fromMap(map['returnType'])
-          : null,
-      pageType: map['pageType'] != null
-          ? ImportableType.fromMap(map['pageType'])
-          : null,
+      returnType: map['returnType'] != null ? ImportableType.fromMap(map['returnType']) : null,
+      pageType: map['pageType'] != null ? ImportableType.fromMap(map['pageType']) : null,
       navigationType: NavigationType.values[map['navigationType']],
-      parameters: List<ImportableType>.from(map['parameters']?.map(
-              (dynamic x) => ImportableType.fromMap(x as Map<String, dynamic>))
-          as Iterable),
-      guards: map['guards'] == null
-          ? null
-          : List<ImportableType>.from(map['guards'].map((dynamic x) =>
-              ImportableType.fromMap(x as Map<String, dynamic>)) as Iterable),
-      defaultValues:
-          map['defaultValues'] as Map<String, dynamic>? ?? <String, dynamic>{},
+      parameters: List<RouteFieldConfig>.from(map['parameters']?.map((dynamic x) => RouteFieldConfig.fromMap(x as Map<String, dynamic>)) as Iterable),
+      guards: map['guards'] == null ? null : List<ImportableType>.from(map['guards'].map((dynamic x) => ImportableType.fromMap(x as Map<String, dynamic>)) as Iterable),
     );
   }
 
   String toJson() => json.encode(toMap());
 
-  factory RouteConfig.fromJson(String source) =>
-      RouteConfig.fromMap(json.decode(source));
+  factory RouteConfig.fromJson(String source) => RouteConfig.fromMap(json.decode(source));
 }