+++ flutter_navigation_generator/test/code_builder/on_generate_route_builder_test.dart
@@ -6,6 +6,7 @@ import 'package:code_builder/code_builder.dart';
 import 'package:flutter_navigation_generator/src/code_builder/on_generate_route_builder.dart';
 import 'package:flutter_navigation_generator/src/models/importable_type.dart';
 import 'package:flutter_navigation_generator/src/models/route_config.dart';
+import 'package:flutter_navigation_generator/src/models/route_field_config.dart';
 import 'package:flutter_navigation_generator_annotations/flutter_navigation_generator_annotations.dart';
 import 'package:test/test.dart';
 
@@ -25,8 +26,7 @@ void main() {
       expect(content.requiredParameters.first.type?.symbol, 'RouteSettings');
       expect(content.requiredParameters.length, 1);
       expect(content.body is Block, true);
-      expect(
-          (content.body as Block).statements.first.toString(), 'return null;');
+      expect((content.body as Block).statements.first.toString(), 'return null;');
       expect((content.body as Block).statements.length, 1);
     });
 
@@ -43,8 +43,7 @@ void main() {
       ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements =
-          (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    settingsUri.queryParameters.forEach((key, value) {\n      arguments[key] ??= value;\n    });",
         "switch (settingsUri.path) {case RouteNames.test: return MaterialPageRoute<void>(builder: (_) => TestPage(), settings: settings, fullscreenDialog: false,);}",
@@ -79,8 +78,7 @@ void main() {
       ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements =
-          (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    settingsUri.queryParameters.forEach((key, value) {\n      arguments[key] ??= value;\n    });",
         "switch (settingsUri.path) {case RouteNames.testWithoutReturn: return MaterialPageRoute<void>(builder: (_) => TestPageWithoutReturn(), settings: settings, fullscreenDialog: false,);case RouteNames.testWithReturn: return MaterialPageRoute<ReturnType>(builder: (_) => TestPageWithReturn(), settings: settings, fullscreenDialog: false,);}",
@@ -118,8 +116,7 @@ void main() {
       ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements =
-          (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    settingsUri.queryParameters.forEach((key, value) {\n      arguments[key] ??= value;\n    });",
         "switch (settingsUri.path) {case RouteNames.testWithoutReturn: return MaterialPageRoute<void>(builder: (_) => TestPageWithoutReturn(), settings: settings, fullscreenDialog: false,);case RouteNames.testWithReturn: return MaterialPageRoute<ReturnType>(builder: (_) => TestPageWithReturn(), settings: settings, fullscreenDialog: false,);}",
@@ -145,8 +142,7 @@ void main() {
       ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements =
-          (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    settingsUri.queryParameters.forEach((key, value) {\n      arguments[key] ??= value;\n    });",
         "switch (settingsUri.path) {case RouteNames.r404: return MaterialPageRoute<void>(builder: (_) => Route404(), settings: settings, fullscreenDialog: false,);}",
@@ -173,8 +169,7 @@ void main() {
       ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements =
-          (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    settingsUri.queryParameters.forEach((key, value) {\n      arguments[key] ??= value;\n    });",
         "switch (settingsUri.path) {case RouteNames.test: return MaterialPageRoute<void>(builder: (_) => TestPage.constructorName(), settings: settings, fullscreenDialog: false,);}",
@@ -200,8 +195,7 @@ void main() {
       ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements =
-          (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    settingsUri.queryParameters.forEach((key, value) {\n      arguments[key] ??= value;\n    });",
         "switch (settingsUri.path) {case RouteNames.test: return MaterialPageRoute<void>(builder: (_) => TestPage(), settings: settings, fullscreenDialog: true,);}",
@@ -230,8 +224,7 @@ void main() {
       ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements =
-          (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    settingsUri.queryParameters.forEach((key, value) {\n      arguments[key] ??= value;\n    });",
         "switch (settingsUri.path) {case RouteNames.test: return CustomPageType<void>(builder: (_) => TestPage(), settings: settings, fullscreenDialog: false,);}",
@@ -260,8 +253,7 @@ void main() {
       ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements =
-          (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    settingsUri.queryParameters.forEach((key, value) {\n      arguments[key] ??= value;\n    });",
         "switch (settingsUri.path) {case RouteNames.test: return CustomPageType<void>(builder: (_) => TestPage(), settings: settings, fullscreenDialog: false,);}",
@@ -283,17 +275,27 @@ void main() {
               className: 'TestPage',
             ),
             parameters: [
-              const ImportableType(
-                  className: 'bool', name: 'testBool', isNullable: true),
-              const ImportableType(className: 'int', name: 'testInt'),
+              RouteFieldConfig(
+                type: const ImportableType(className: 'bool', name: 'testBool', isNullable: true),
+                addToJson: true,
+                defaultValue: null,
+                ignore: false,
+                queryName: 'testBool',
+              ),
+              RouteFieldConfig(
+                type: const ImportableType(className: 'int', name: 'testInt'),
+                addToJson: true,
+                defaultValue: null,
+                ignore: false,
+                queryName: 'testInt',
+              ),
             ],
           ),
         },
       ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements =
-          (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    settingsUri.queryParameters.forEach((key, value) {\n      arguments[key] ??= value;\n    });",
         "switch (settingsUri.path) {case RouteNames.test: return MaterialPageRoute<void>(builder: (_) => TestPage(testBool: arguments['testBool'] is String ? arguments['testBool'] == 'true' : arguments['testBool'] as bool?,testInt: arguments['testInt'] is String ? int.parse(arguments['testInt']) : arguments['testInt'] as int,), settings: settings, fullscreenDialog: false,);}",
@@ -315,17 +317,27 @@ void main() {
               className: 'TestPage',
             ),
             parameters: [
-              const ImportableType(
-                  className: 'bool', name: 'testBool', isNullable: true),
-              const ImportableType(className: 'int', name: 'testInt'),
+              RouteFieldConfig(
+                type: const ImportableType(className: 'bool', name: 'testBool', isNullable: true),
+                addToJson: true,
+                defaultValue: null,
+                ignore: false,
+                queryName: 'testBool',
+              ),
+              RouteFieldConfig(
+                type: const ImportableType(className: 'int', name: 'testInt'),
+                addToJson: true,
+                defaultValue: null,
+                ignore: false,
+                queryName: 'testInt',
+              ),
             ],
           ),
         },
       ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements =
-          (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    settingsUri.queryParameters.forEach((key, value) {\n      arguments[key] ??= value;\n    });",
         "switch (settingsUri.path) {}",
@@ -350,17 +362,27 @@ void main() {
               className: 'TestPage',
             ),
             parameters: [
-              const ImportableType(
-                  className: 'bool', name: 'testBool', isNullable: true),
-              const ImportableType(className: 'int', name: 'testInt'),
+              RouteFieldConfig(
+                type: const ImportableType(className: 'bool', name: 'testBool', isNullable: true),
+                addToJson: true,
+                defaultValue: null,
+                ignore: false,
+                queryName: 'testBool',
+              ),
+              RouteFieldConfig(
+                type: const ImportableType(className: 'int', name: 'testInt'),
+                addToJson: true,
+                defaultValue: null,
+                ignore: false,
+                queryName: 'testInt',
+              ),
             ],
           ),
         },
       ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements =
-          (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    settingsUri.queryParameters.forEach((key, value) {\n      arguments[key] ??= value;\n    });",
         "switch (settingsUri.path) {case RouteNames.testPage: return MaterialPageRoute<void>(builder: (_) => TestPage(testBool: arguments['testBool'] is String ? arguments['testBool'] == 'true' : arguments['testBool'] as bool?,testInt: arguments['testInt'] is String ? int.parse(arguments['testInt']) : arguments['testInt'] as int,), settings: settings, fullscreenDialog: false,);}",
@@ -382,17 +404,27 @@ void main() {
               className: 'TestPage',
             ),
             parameters: [
-              const ImportableType(
-                  className: 'bool', name: 'testBool', isNullable: true),
-              const ImportableType(className: 'int', name: 'testInt'),
+              RouteFieldConfig(
+                type: const ImportableType(className: 'bool', name: 'testBool', isNullable: true),
+                addToJson: true,
+                defaultValue: null,
+                ignore: false,
+                queryName: 'testBool',
+              ),
+              RouteFieldConfig(
+                type: const ImportableType(className: 'int', name: 'testInt'),
+                addToJson: true,
+                defaultValue: null,
+                ignore: false,
+                queryName: 'testInt',
+              ),
             ],
           ),
         },
       ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements =
-          (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    settingsUri.queryParameters.forEach((key, value) {\n      arguments[key] ??= value;\n    });",
         "switch (settingsUri.path) {}",
@@ -416,21 +448,27 @@ void main() {
               className: 'TestPage',
             ),
             parameters: [
-              const ImportableType(
-                  className: 'bool', name: 'testBool', isNullable: true),
-              const ImportableType(className: 'int', name: 'testInt'),
+              RouteFieldConfig(
+                type: const ImportableType(className: 'bool', name: 'testBool', isNullable: true),
+                addToJson: true,
+                defaultValue: 'true',
+                ignore: false,
+                queryName: 'testBool',
+              ),
+              RouteFieldConfig(
+                type: const ImportableType(className: 'int', name: 'testInt'),
+                addToJson: true,
+                defaultValue: '1',
+                ignore: false,
+                queryName: 'testInt',
+              ),
             ],
-            defaultValues: {
-              'testBool': 'true',
-              'testInt': '1',
-            },
           ),
         },
       ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements =
-          (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    settingsUri.queryParameters.forEach((key, value) {\n      arguments[key] ??= value;\n    });",
         "switch (settingsUri.path) {case RouteNames.test: return MaterialPageRoute<void>(builder: (_) => TestPage(testBool: arguments['testBool'] is String ? arguments['testBool'] == 'true' : arguments['testBool'] as bool?,testInt: arguments['testInt'] is String ? int.parse(arguments['testInt']) : arguments['testInt'] as int,), settings: settings, fullscreenDialog: false,);}",
@@ -457,8 +495,7 @@ void main() {
       ).generate();
 
       expect(content.body is Block, true);
-      expect(
-          (content.body as Block).statements.first.toString(), 'return null;');
+      expect((content.body as Block).statements.first.toString(), 'return null;');
       expect((content.body as Block).statements.length, 1);
     });
 
@@ -476,8 +513,7 @@ void main() {
       ).generate();
 
       expect(content.body is Block, true);
-      expect(
-          (content.body as Block).statements.first.toString(), 'return null;');
+      expect((content.body as Block).statements.first.toString(), 'return null;');
       expect((content.body as Block).statements.length, 1);
     });
 
@@ -495,8 +531,7 @@ void main() {
       ).generate();
 
       expect(content.body is Block, true);
-      expect(
-          (content.body as Block).statements.first.toString(), 'return null;');
+      expect((content.body as Block).statements.first.toString(), 'return null;');
       expect((content.body as Block).statements.length, 1);
     });
 
@@ -522,8 +557,7 @@ void main() {
       ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements =
-          (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    settingsUri.queryParameters.forEach((key, value) {\n      arguments[key] ??= value;\n    });",
         "switch (settingsUri.path) {case RouteNames.test: final testGuard = guards.whereType<TestGuard>().first;\nif (!testGuard.value) {\n  guardedRouteSettings = settings;\n  return onGenerateRoute(RouteSettings(\n    arguments: settings.arguments,\n    name: testGuard.alternativeRoute,\n  ));\n}\nreturn MaterialPageRoute<void>(builder: (_) => TestPage(), settings: settings, fullscreenDialog: false,);case RouteNames.test2: return MaterialPageRoute<void>(builder: (_) => TestPage2(), settings: settings, fullscreenDialog: false,);}",
@@ -559,8 +593,7 @@ void main() {
       ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements =
-          (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    settingsUri.queryParameters.forEach((key, value) {\n      arguments[key] ??= value;\n    });",
         "switch (settingsUri.path) {case RouteNames.test: final testGuard = guards.whereType<TestGuard>().first;\nif (!testGuard.value) {\n  guardedRouteSettings = settings;\n  return onGenerateRoute(RouteSettings(\n    arguments: settings.arguments,\n    name: testGuard.alternativeRoute,\n  ));\n}\nreturn MaterialPageRoute<void>(builder: (_) => TestPage(), settings: settings, fullscreenDialog: false,);case RouteNames.test2: return MaterialPageRoute<void>(builder: (_) => TestPage2(), settings: settings, fullscreenDialog: false,);}",