+++ flutter_navigation_generator/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 2.1.0 (2024-12-06)
+
+- Added `@FlutterRouteField` annotation to specify fields with the following options:
+    - `ignore`: Defaults to false (except for Key? key field), excludes this field from the goTo method
+    - `addToJson`: default true, add this field to the json serialization, only useful for web, use this to add cached values, but prevent the user from overriding them
+    - `queryName`: default null, use this to specify the query parameter name, if not specified, the field name is used
+
 ## 2.0.3 (2024-10-21)
 
 - Fixed a nullability issue